Allow preflight requests

+++ app/src/Controller/ApiController.php
@@ -31,6 +31,10 @@ class ApiController extends Controller {
   }
 
   public function new(Request $request) {
+    if ($request->getMethod() == 'OPTIONS') {
+      return new Response(Response::HTTP_OK);
+    }
+
     $body = json_decode($request->getContent(), true);
     $apiResponse = $this->api->new(self::DEFAULT_SIZE, $body['mode'] === 'multi');
     $this->em->persist($apiResponse['player']);
@@ -43,7 +47,11 @@ class ApiController extends Controller {
     ]);
   }
 
-  public function game(GameContext $context) {
+  public function game(Request $request, GameContext $context) {
+    if ($request->getMethod() == 'OPTIONS') {
+      return new Response(Response::HTTP_OK);
+    }
+
     $game = $context->getGame();
     $currentPlayer = $context->getPlayer();
     $player = $currentPlayer ?: $game->getPlayer1();
@@ -71,7 +79,11 @@ class ApiController extends Controller {
     ]);
   }
 
-  public function cancel(GameContext $context) {
+  public function cancel(Request $request, GameContext $context) {
+    if ($request->getMethod() == 'OPTIONS') {
+      return new Response(Response::HTTP_OK);
+    }
+
     if ($context->getIsPlayer()) {
       $this->em->remove($context->getGame());
       $this->em->flush();
@@ -80,12 +92,14 @@ class ApiController extends Controller {
         Response::HTTP_OK
       );
     }
-    return new Response(
-      Response::HTTP_INTERNAL_SERVER_ERROR
-    );
+    return new Response(Response::HTTP_INTERNAL_SERVER_ERROR);
   }
 
-  public function join(GameContext $context, TokenGenerator $tokenGenerator) {
+  public function join(Request $request, GameContext $context, TokenGenerator $tokenGenerator) {
+    if ($request->getMethod() == 'OPTIONS') {
+      return new Response(Response::HTTP_OK);
+    }
+
     $game = $context->getGame();
 
     if ($game->isFull()) {
@@ -110,7 +124,11 @@ class ApiController extends Controller {
     ]);
   }
 
-  public function move(GameContext $context, int $tile) {
+  public function move(Request $request, GameContext $context, int $tile) {
+    if ($request->getMethod() == 'OPTIONS') {
+      return new Response(Response::HTTP_OK);
+    }
+
     $game = $context->getGame();
     $currentPlayer = $context->getPlayer();
 
@@ -128,7 +146,11 @@ class ApiController extends Controller {
     ]);
   }
 
-  public function games() {
+  public function games(Request $request) {
+    if ($request->getMethod() == 'OPTIONS') {
+      return new Response(Response::HTTP_OK);
+    }
+
     $gameIds = $this->gameRepository->findOpenMultiplayerGames();
 
     return new JsonResponse([
